requirements.txt
pydantic>=2.0.0
openai>=1.35.0
fastapi>=0.116.0
uvicorn>=0.30.0
pyyaml>=6.0
httpx>=0.27.0
